Add re-check logic when create network failed with conflict
Add re-check logic when create network faild with the Conflict Exception
whose status_code is 409, it's means another task of creating the same
network is happend and done. So there is no need to create it again, instead
we query if the network exists(by name). If we can not find the network,
then the exception raised.

Change-Id: I44a6909b3119be748d4445a0f560454c19687276