Periodic update under aggregate/resources info for schedule (part 1)
Before, the cascading don't know the aggregates and resource info of
the cascadeds managed by proxys, so we cannot using the most filters
to choose the proper host with nova scheduler.

By updating the aggregates(includes hosts and metadata) and resource
used info in cascaded to the cascading periodically, this problem can
be solved.

Change-Id: If369f0023130c24544bbdbb3c1afbef05527d783