Implement security group - Neutron plugin part
The task Neutron plugin needs to finish for security group
functionality is simple. When a rule is created or deleted
in a security group, it just updates the corresponding
bottom security group. Also, creating or deleting a rule
with "remote_group_id and changing rules in default
security group will be rejected by plugin.

Currently this task runs in synchronous way. Later it can
be implemented in asynchronous way for better response time.

Change-Id: Ibbf46c2e91382986c02324d86bc22887e93267eb