Manual sync from global-requirements
In review I94f6f5f853078feeccaea0c50e5690101f95e318 tricircle is asking
to be added to projects.txt.  Doing so will mean that when
global-requirements.txt is altered in a away that directly impacts
tricircle an automated (by proposal-bot) review similar to this one will
be generated.

This review manually performs this sync so that the first bot generated
change will be smaller and easier to review.

This introduces several significant changes (eventlet, ovo, Babel) so
please test carefully.

Change-Id: Ib2b7ed2671a5ad8d2a6262e07e0bbc5f921abf33