Update devstack configure in stable/mitaka branch
1.What is the problem:
The devstack configuration is still pointed to the master branch in
stable/mitaka branch

2.What needs to be fixed:
In the Tricircle stable/mitaka branch, the devstack configuration should
be changed to downlowd the code from the Tricircle stable/mitaka branch
but not from the master branch. And also use Mitaka Neutron release in
running devstack for version compatibility purpose but instead of the Neutron
master branch.

3.What is the purpose of this patch set:
The Tricircle stable/mitaka branch has no enough features to run the
Tricircle in production, but after this patch set is merged, we can make a
tag in this branch, and check whether the release publishing process works
or not. The release package publishing to PyPI job has been configured and
merged in the following patch set: https://review.openstack.org/#/c/323108/

Change-Id: Ic273eaa033dfb53fcd17e9e91945341bb9f88e58
Signed-off-by: Chaoyi Huang <joehuang@huawei.com>