Adapt error response body in nova and cinder api
1. What is the problem
In the current controller implementation in nova_apigw and
cinder_apigw, pecan.abort is used to raise an exception when
error occurs. The problem of using pecan.abort is that the
response body doesn't have the same format with the error
response body in nova api and cinder api. Thus python client
may not correctly extract the error message, also, tempest
test may fail.

2. What is the solution to the problem
Replace pecan.abort with correct response body.

3. What the features need to be implemented to the Tricircle
   to realize the solution
In this patch, we remove pecan.abort calls in controllers
of nova and cinder resources and directly return the error
response body with correct format. Controllers for the Tricircle
api still keep pecan.abort calls since we don't have special
requirement on the format of error response body.

Change-Id: I0e6fe9ddfce3f001fee0be2160d24c9c628d0a88