Add network type support to the Tricircle plugin
1. What is the problem
In the current implementation of the Tricircle plugin for neutron,
network type is not supported so users cannot create networks
with network type specified. In the specification of cross-pod
l2 networking feature[1], we decide to support several network
types like local, shared VLAN, shared VxLAN, etc, the first step
is to make the Tricircle plugin be aware of network type.

2. What is the solution to the problem
Handle network type in the Tricircle plugin for neutron.

3. What the features need to be implemented to the Tricircle
   to realize the solution
In this patch, we add a framework to load type driver which
processes different network type. The framework is based on
neutron ML2 implemenation, we inherit the ML2 type manager and
create a new Tricircle type manager. Also the Tricircle plugin
is modified to extract network type parameter from request and
insert network type information to response.

[1] https://github.com/openstack/tricircle/blob/master/specs/cross-pod-l2-networking.rst

Change-Id: Ida9b88df6113db46e637a7841ce5c1adaf651eba