Fix issue from _add_network_segment in Neutron
1. What is the problem
The Tricircle Neutron plugin will call Neutron type manager's
function _add_network_segement, but Neutron has removed the
"mtu" parameter recently, this makes the check and gate test
in the Tricircle failed because the Tricircle still passes the
"mtu" parameter.

All check and gate test of python27 failed due to this change,
and any new patch is not able to be merged. The issue should
be fixed ASAP.

2. What's needed to be fixed:
Do not pass "mtu" to the function _add_network_segement.

3. What is the purpose of this patch set:
To get rid of the failure in check and gate test failure.

Change-Id: Id51cc9840e4bf2dd8e01b504502266e962c1e0c9