Central and local plugin (part 1, base)
1. What is the problem
As discussed in the feature specification[1], we are going to
move the process of networking automation from the Nova-APIGW
to Neutron server.

2. What is the solution to the problem
Implement a new Neutron core plugin which runs in local Neutron
server to finish the networking automation process. Also, the
original Neutron core plugin is renamed as central plugin and
needs some changes to work with local plugin.

3. What the features need to be implemented to the Tricircle
   to realize the solution
With this patch, users can boot a virtual machine directly via
the local Nova server. But security group support is not covered.

DevStack script and local.conf sample are also updated.

[1] https://github.com/openstack/tricircle/blob/master/specs/ocata/local-neutron-plugin.rst

Change-Id: I6a3dc5e9af395e3035a7d218264a08b6313a248d