Fix the link error in the document README.rst
1. What is the problem?
The Tricircle Admin API documentation has been implemented, so the
relevant link to it in the document README.rst should be updated as
well.

2. What is the solution to the problem?
Remove the original null link at the bottom of the document README.rst and
change the part "Documentation" to "Tricircle Admin API documentation", then
provide the relative link behind it.

3. What the features need to be implemented to the Tricircle to realize the solution?
None.

Change-Id: I8dae02ddef4a3019abe91ec1ca283280b64c7c61