Specification for local Neutron plugin
1. What is the problem
Part of the networking automation functionality is implemented
in the Nova-APIGW, which bring coupling between Neutron and Nova
servers. Users must deploy Nova-APIGW if they want to use the
networking automation function, which is not flexible.

2. What is the solution to the problem
Move the process of networking automation from the Nova-APIGW
to local Neutron plugin. This specification covers the detailed
introduction and discussion of this plugin.

3. What the features need to be implemented to the Tricircle
   to realize the solution
The local Neutron plugin discussed in this specification needs
to be implemented.

Change-Id: I6450ab0d2103eb8fd6fa8869ae80376a37ae371a