Central and local plugin (part 2, l3 functionality)
1. What is the problem
Necessary changes for local plugin and central plugin to boot
a virtual machine have been submitted in this patch[1]. As the
next step, we need to add l3 functionality to the local and
central plugin.

2. What is the solution to the problem
Several changes in local plugin.
(1) Before creating local subnet, local plugin sends request to
central plugin to create a "reserved gateway port", and use the
ip address of this port as the gateway ip of the local subnet.
(2) When local plugin receives network or subnet creation request,
if the request contains "name" parameter and the name is a UUID,
local plugin uses the name as the id of the local network or
subnet.

3. What the features need to be implemented to the Tricircle
   to realize the solution
With this patch, users can connect virtual machines booted
directly via the local Nova server in different networks with
a router.

[1] https://review.openstack.org/375281

Change-Id: I12094f30804c0bad2f74e0ff510ac26bd217cfd4