Implement resource routing features
1. What is the problem?
In order to dispatch resource operation request to the proper bottom instance,
we need a resource routing table, which maps a resource to the bottom instance
which it’s located at. However, the resource routing APIs have not been
implemented, so this patch carries out these features and performs some tests
for them at the same time.

2. What is the solution to the problem?
To implement the resource routing features, a new file named routing.py is
created in the directory tricircle/api/controllers, then regular post and
get functions will be implemented there. As for the unit test, a file named
test_routing.py is added to the directory tricircle/tests/unit/api/controllers,
conducting relevant tests for the features in file routing.py.

3. What the features need to be implemented to the Tricircle to realize
the solution?
Some main features like post, update, get and delete functions are implemented
in the file routing.py, then relative test method for each of them is realized
in the file test_routing.py. At last, these features are added to the search path
in the file root.py for REST API call.

Change-Id: I2c1073b8c8bd17eeb395afaa5ab7ea8d63609222