Update the dependency in stable/newton branch
1. What is the problem
The Tricircle is one project in OpenStack name space, it runs with the
dependency to OpenStack services like Nova, Cinder, Neutron. In the
master branch of the Tricircle, all dependencies are pointed to the
master branch of all OpenStack services, include itself.

When the stable/newton branch of the Tricircle was created, all these
dependencies are not changed, for the master will change now and then,
then the stable/newton branch of the Tricircle will not be able to run
normally as more and more changes happen in the master branch, for
the Tricircle must be changed too to adapt the dependency change, but in
the stable/newton branch, we'll not cherry pick back all these changes
from master to the stable/newton branch.

2. What is the solution to the problem
When the Tricircle stable/newton branch is created, all these
dependencies should be pointed to stable/newton branch.

3. What the features need to be implemented to the Tricircle
   to realize the solution
No new feature is needed, just update the dependency.

Change-Id: I2c63c5026fbc3d1796f9b2a18a5476d575ae592f
Signed-off-by: joehuang <joehuang@huawei.com>