Update multi-node installation guide with DevStack
1. What is the problem?
The current installation guide is based on the Tricircle with
API-GW and networking automation functionality. After the Tricircle
splitting, only networking automation will be remained in the
Tricircle repository, so the current installation guide will not
work anymore.

2. What is the solution to the problem?
Update the installation guide to reflect the Tricricle splitting.
This patch will focus on multi-node installation with DevStack.
Single node installation with DevStack has been updated in this
patch[1].

3. What the features need to be implemented to the Tricircle to realize
the solution?
No new features.

[1] https://review.openstack.org/#/c/384872

Change-Id: I8ee326024fd152133829caced69a68348172f065