Add manual installation guide
1. What is the problem?
We already have guides for users to install the Tricircle with DevStack.
But a guide discussing how to install the Tricircle step by step without
DevStack is also needed for users who install OpenStack manually.

2. What is the solution to the problem?
Add manual installation guide, and link it to the Tricircle docs page.

3. What the features need to be implemented to the Tricircle to realize
the solution?
No new features

Change-Id: I45344df2ccb27c5096d66f00e4e76efa07d66012