Layer-3 networking and combined bridge network spec
1. What is the problem
The current implementation of bridge networks has some problems
when supporting DVR and shared VxLan network.

2. What is the solution to the problem
The detailed discussion of the problems and the proposals to
solve the problems will be covered in this specification. This
specification will also discuss the proposed and supported
cross-OpenStack layer-3 networking model.

3. What the features need to be implemented to the Tricircle
   to realize the solution
The proposal to combine bridge networks discussed in this
specification needs to be implemented.

Change-Id: Ib0c3d8c31de5c2f9c1c5b08a222971ea628cc3c1