Documentation on resource routing API
1. What is the problem?
The resource routing API has been completed, but the relevant documentation
is not there.

2. What is the solution to the problem?
The resource routing API is part of the Tricircle Admin API, so the
documentation about it is added to the tail of the file
/tricircle/doc/source/api_v1.rst. For resource routing, there are five main
public functions. This patch contains explanations to each function and the
request and response examples of them.

3. What the features need to be implemented to the Tricircle to realize the
solution?
No new features.

Change-Id: I7ef48d071f4c49455a7eabb284839047943bbd6a